+++ backend/garpix_keycloak/mixins.py
@@ -3,9 +3,23 @@ from django.utils.translation import gettext_lazy as _
 from garpix_utils.string import get_random_string
 from django.db.models import Q
 
+from garpix_keycloak.models import KeycloakGroup
+
 
 class KeycloakUserMixin(models.Model):
     keycloak_id = models.CharField(max_length=128, null=True, blank=True, verbose_name=_('Keycloak ID'))
+
+    groups = models.ManyToManyField(
+        KeycloakGroup,
+        verbose_name=_('keycloak groups'),
+        blank=True,
+        help_text=_(
+            'The keycloak groups this user belongs to. A user will get all permissions '
+            'granted to each of their groups.'
+        ),
+        related_name="user_set",
+        related_query_name="user",
+    )
 
     class Meta:
         abstract = True